Remove the enrollmentStorage and integrate with LedgerStorage
DB Instance is managed in one class and the required query operations are managed in LedgerStorage.
